identity tests (#109)

+++ be-identity/identity/orchestrations.py
@@ -10,15 +10,24 @@ import common.datatypes.exception
 import identity.datatypes.request
 import identity.datatypes.response
 import identity.services.auth
+import identity.services.user
 
 
 DEFAULT_ADMIN_EMAIL = os.environ.get("DEFAULT_ADMIN_EMAIL")
 
 
 async def arad_register(
-    email: str, passphrase: str, database: sqlmodel.Session
+    email: str,
+    passphrase: str,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.RegisterResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
+
     try:
         user = await auth_service.create_user_with_passphrase(
             email=email,
@@ -29,9 +38,6 @@ async def arad_register(
 
     if DEFAULT_ADMIN_EMAIL and email == DEFAULT_ADMIN_EMAIL:
         await auth_service.assign_role_for_user_id(
-            user_id=user.id, role=common.datatypes.domain.Role.REVIEWER
-        )
-        await auth_service.assign_role_for_user_id(
             user_id=user.id, role=common.datatypes.domain.Role.ADMINISTRATOR
         )
 
@@ -40,14 +46,21 @@ async def arad_register(
     return identity.datatypes.response.RegisterResponse(
         refresh_token=refresh_token,
         user=user,
-        roles=user.roles,
     )
 
 
 async def arad_login(
-    email: str, passphrase: str, database: sqlmodel.Session
+    email: str,
+    passphrase: str,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.LoginResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
+
     try:
         user = await auth_service.authenticate_user_by_email_and_passphrase(
             email=email,
@@ -61,36 +74,58 @@ async def arad_login(
     return identity.datatypes.response.LoginResponse(
         refresh_token=refresh_token,
         user=user,
-        roles=user.roles,
     )
 
 
 async def arad_logout(
-    refresh_token: str, database: sqlmodel.Session
+    refresh_token: str,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.LogoutResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
+
     await auth_service.destroy_refresh_token(refresh_token=refresh_token)
 
     return identity.datatypes.response.LogoutResponse(status="ok")
 
 
 async def arad_access_token(
-    refresh_token: str, scope: common.datatypes.domain.Role, database: sqlmodel.Session
+    refresh_token: str,
+    scope: common.datatypes.domain.Role,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.TokenResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
-    user_id = await auth_service.verify_and_extract_uuid_from_refresh_token(
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
+
+    user_id = await auth_service.verify_and_extract_user_id_from_refresh_token(
         refresh_token=refresh_token
     )
 
-    access_token = await auth_service.create_access_token(user_id=user_id, scope=scope)
+    access_token = await auth_service.verify_role_and_create_access_token(
+        user_id=user_id, scope=scope
+    )
 
     return identity.datatypes.response.TokenResponse(access_token=access_token)
 
 
 async def arad_roles(
-    database: sqlmodel.Session,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.RolesResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
+
     roles = await auth_service.all_roles()
 
     return identity.datatypes.response.RolesResponse(roles=roles)
@@ -100,9 +135,14 @@ async def arad_modify_role_assignment(
     user_id: uuid.UUID,
     role: common.datatypes.domain.Role,
     action: identity.datatypes.request.RoleAction,
-    database: sqlmodel.Session,
+    auth_service: identity.services.auth.AuthService | None = None,
+    database: sqlmodel.Session | None = None,
 ) -> identity.datatypes.response.RoleResponse:
-    auth_service = identity.services.auth.AuthService(database=database)
+    if auth_service is None and database is None:
+        raise Exception()
+
+    if auth_service is None:
+        auth_service = identity.services.auth.AuthService(database=database)
 
     # this admin function isn't critical and the ui should protect us from most edge cases so we'll just let these
     # calls explode if for instance the role has already been assigned to the user (possible with two tabs open)
@@ -113,7 +153,23 @@ async def arad_modify_role_assignment(
     else:
         raise Exception()
 
-    auth_service = identity.services.auth.AuthService(database=database)
     await auth_service.destroy_all_refresh_tokens_for_user_id(user_id=user_id)
 
     return identity.datatypes.response.RoleResponse(role=role)
+
+
+async def arad_users(
+    email_filter: str,
+    page: int | None,
+    database: sqlmodel.Session | None,
+    user_service: identity.services.user.UserService | None,
+) -> identity.datatypes.response.UsersResponse:
+    if user_service is None and database is None:
+        raise Exception()
+
+    if user_service is None:
+        user_service = identity.services.user.UserService(database=database)
+
+    user_page = await user_service.page(email_filter=email_filter, number=page)
+
+    return identity.datatypes.response.UsersResponse(**user_page.dict())
